pyproject.toml
[tool.poetry]
name = "autom8-calendar-integration"
version = "0.1.0"
description = ""
authors = ["hermansen <hermansendev@gmail.com>"]

[tool.poetry.scripts]
init = "autom8_calendar_integration:start"

[tool.poetry.dependencies]
python = "^3.8"
mypy = "^0.761"
google-api-python-client = "^1.7.11"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
pendulum = "^2.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
